requirements.txt
pandas>=2.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
